refactor: move pusher config to workflow/schema/ — decouple from task
Introduces michelangelo/workflow/schema/ as the canonical home for task
configuration dataclasses, mirroring the pattern already established in
the internal uber-one codebase (schema/v2alpha1/).

Key changes:
- workflow/schema/exceptions.py: ConfigurationError defined here
  (schema-layer exception, not a runtime PusherError subclass)
- workflow/schema/pusher.py: canonical home for all pusher config classes
  (DatasetFormat, ModelPluginConfig, DatasetPluginConfig,
  EvalReportPluginConfig, PusherPluginConfig, PusherConfig)
- workflow/tasks/pusher/config.py: re-export shim for backwards compat
- workflow/tasks/pusher/exceptions.py: re-exports ConfigurationError
  from schema layer; PusherError and runtime errors remain here
- workflow/tasks/pusher/registry.py: import ConfigurationError from
  schema layer directly
- Tests updated: test_config.py imports from schema canonical path;
  test_exceptions.py reflects ConfigurationError is no longer a
  PusherError subclass

This allows workflow definitions to import PusherConfig directly from
schema/ without coupling to the pusher task package.
