+++ python/michelangelo/workflow/schema/__init__.py
@@ -1,0 +1,1 @@
+"""Schema package for workflow task configuration dataclasses."""
